% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour.R
\name{create}
\alias{create}
\title{Create a Vapour package}
\usage{
create(path, ...)
}
\arguments{
\item{path}{Path where to create the package.}

\item{...}{Passed to \code{\link[usethis:create_package]{usethis::create_package()}}.}
}
\description{
A light wrapper around \code{\link[usethis:create_package]{usethis::create_package()}}
to create an R package and \code{\link[=scaffold]{scaffold()}}.
}
